feat: translate signature.gd to C# and rename
This script was located on the sprite instead of the scene's root for some reason, so to be consistent with other scenes, I moved it up to the scene root and changed references accordingly. It also did extra unneccesary calls to set the modulate property of nodes so I removed those.